refactor(workersvc): move input submission/approval to submit.go
Pure same-package code motion (PRD #921 M2, seam 3). Moves the 13 submit/
approval functions (SubmitInput, SubmitInputWithCapabilityOverride,
submitInput, enqueueRunInput, submitApproval, capabilityGate,
effectiveOwningWorkerCaps, OverrideRunRequiredCapabilities, submitAnswer,
orEmpty, clampInt, submitScopeCeiling, stopKindFor) out of service.go into
the new submit.go, byte-identically. stopKindFor moves here because its only
callers are inside SubmitInput. The two inline //nolint:gosec directives and
the load-bearing "workersvc: approve-time milestone freeze" log string travel
verbatim. The now-unused "strconv" import is dropped from service.go (pure
deletion; its only users moved). hasLivePoller stays. No exported-API change,
no behavior change; go doc output unchanged.
